Refactor creation of domain_map into confdb
The NSS provider, the Data Provider backends and the InfoPipe all
need access to the domain map provided by the confdb. Instead of
reimplimenting it in multiple places, it is now provided in a pair
of helper functions from the confdb.

confdb_get_domains() returns a domain map by reference. Always
returns the most up-to-date set of domains from the confdb.

confdb_get_domains_list() returns an array of strings of all the
domain names. Always returns the most up-to-date set of domains
from the confdb.

This patch also modifies the btreemap_get_keys() function to
better handle memory and report allocation failures.