Implement CreateUser in InfoPipe
Changed the order of the arguments to CreateUser in the
Introspection XML to match the other functions (domain belongs
second on the list)

A few other minor fixes as well:
Fixed a typo in SYSDB_GETCACHED_FILTER and sysdb_transaction_end().

Added missing error handling in infp_do_user_set_uid().