Fix race condition with initial sysdb creation
When the sysdb LDB file does not exist on the system, the first
attempt to connect to it will invoke a creation routine. However,
both the NSS and the InfoPipe are started in parallel by the
monitor, resulting in a race condition as they both try to
initialize the sysdb. The easiest fix for this is to simply have
the monitor create the sysdb before it launches NSS and InfoPipe.