Fix SIGSEGV in InfoPipe startup
If the user that starts InfoPipe is not permitted by the system
bus to request the InfoPipe name, the sssd_info process would
segfault, since the destructor for the connection object was
called before it was completely created. I have moved the
initialization of the destructor to later in the setup routine.