Fix calling setsid and resolve the sssd signal bug
For some reason we were not testing for HAVE_SETSID in configure therefore the
setsid() function was never called. This failed to set the process group after
the first fork.
Remove ifdef because we depend on setsid() anyway, so if it is not available on
some platform it is better to fail rather then silently succeed but not have the
right process group set up.