fix for pam proxy chauthtok
When a user from a domain served by the proxy backend changes his
password with passwd the passwd command asks for the old password,
but it is not validated by the pam_chauthtok call in the proxy
backend, because it is running as root.

If the request is coming the unpriviledged socket we now call
pam_authenticate explicitly before pam_chauthtok.