Allow the use of custom CFLAGS on the make command line
Setting CFLAGS explicitly in configure.ac means that they would be
overwritten when using e.g. make CFLAGS="-O0 -g"
This replaces the explicit setting of CFLAGS with an
AM_CONDITIONAL to have Makefile.am set these instead.

Also fixes a missing #include that was coincidentally obscured
because gcc's -O2 happened to be able to locate it. Setting -O0
revealed the problem.