Monitor resolv.conf for changes
This patch updates the monitor_config_file() functions so that
they can monitor any number of files and invoke a specified
callback whenever they are modified.

When inotify is available, we will add an additional watch
descriptor to the inotify file descriptor.

When inotify is not available, the polling function will simply
loop to check each file in the monitor list.

When changes are discovered in resolv.conf, the monitor will send
a "resInit" signal to all of its known children. They are only
required to handle this function if they need updated DNS
information. Services that do not implement resInit should return
DBUS_ERROR_UNKNOWN_METHOD (rather than timing out) with no ill
effects.