FORMATTING - minor cleanup of the unit test.
This patch adds formatting changes to
better follow the style guidelines in
the collection unit test.
No logical changes to the code.
I was planning to do it for a while
per Simo's comment when he accepted
the core of the collection code but
indicated that the unit test should be
cleaned later.
Later has come.