Eliminate the --with-tests configure flag
--with-tests was confusing. Since we now build our tests only with
'make check', it doesn't make sense for this to be a configure-
time option. We will detect during configure whether the 'check'
package is available and we will use them if so. Otherwise, we
will only build and execute any test suites that do not rely on
the 'check' framework.

We will print warning during 'configure' if CHECK is not installed