Send debug messages to logfile
Introduces a new option --debug-to-files which makes SSSD output its
debug information to a file instead of stderr, which is still the
default.

Also introduces a new confdb option debug_to_files which does the same,
but can be specified per-service in the config file.

The logfiles are stored in /var/log/sssd by default.

Changes the initscript to log to files by default.