INI: Added method to get string list with empty values
The original implementation was compressing the list,
throwing away empty strings.
The function that did that was pretty brain damaging.
I cleaned it up and adjusted so that it could return
list with empty values and without them.
The old function was turned into a wrapper and a new
high level function was intorduced to provide
ability to get both empty and non empty strings.