Make resolve and failover test work with CK_FORK=no
The leak checking code which is used by the resolve and failover tests
frees talloc's autofree context which is not recommended. As a
consequence the initialization of tevent failed when it was called by
the second test and CF_FORK=no, because it holds some data in the
autofree context.

This patch introduces a global talloc context which should be uses by
the test as the root of their memory hierarchy instead of NULL. This
global context is used in the leak checking routines.

Not all types of memory leaks can be detected by the new version , it is
recommended to use valgrind or similar tools additionally.