Make PAM responses more compatible with D-BUS spec
Previously, the PAM responses could contain an arbitrary number of
arguments. This is not acceptable by the D-BUS protocol, as there
is no way to introspect it. This patch converts the response
objects to be an array of D-BUS structs.

It also fixes two potential memory leaks by not unref'ing the
reply object if we get an error.